Update kart characteristics with python script

+++ src/karts/abstract_characteristic.hpp
@@ -312,8 +312,8 @@ public:
     float getAnvilSpeedFactor() const;
 
     float getParachuteFriction() const;
-    int   getParachuteDuration() const;
-    int   getParachuteDurationOther() const;
+    float getParachuteDuration() const;
+    float getParachuteDurationOther() const;
     float getParachuteDurationRankMult() const;
     float getParachuteDurationSpeedMult() const;
     float getParachuteLboundFraction() const;
